+++ include/lmms_math.h
@@ -136,15 +136,15 @@ inline float signedPowf(float v, float e)
 //! Value should be within [0,1]
 inline float logToLinearScale(float min, float max, float value)
 {
-	if( min < 0 )
+	if (min < 0)
 	{
 		const float mmax = std::max(std::abs(min), std::abs(max));
-		const float val = value * ( max - min ) + min;
+		const float val = value * (max - min) + min;
 		float result = signedPowf(val / mmax, numbers::e_v<float>) * mmax;
-		return std::isnan( result ) ? 0 : result;
+		return std::isnan(result) ? 0 : result;
 	}
-	float result = std::powf(value, numbers::e_v<float>) * (max - min) + min;
-	return std::isnan( result ) ? 0 : result;
+	float result = std::pow(value, numbers::e_v<float>) * (max - min) + min;
+	return std::isnan(result) ? 0 : result;
 }
 
 
@@ -152,15 +152,15 @@ inline float logToLinearScale(float min, float max, float value)
 inline float linearToLogScale(float min, float max, float value)
 {
 	const float valueLimited = std::clamp(value, min, max);
-	const float val = ( valueLimited - min ) / ( max - min );
-	if( min < 0 )
+	const float val = (valueLimited - min) / (max - min);
+	if (min < 0)
 	{
 		const float mmax = std::max(std::abs(min), std::abs(max));
 		float result = signedPowf(valueLimited / mmax, numbers::inv_e_v<float>) * mmax;
-		return std::isnan( result ) ? 0 : result;
+		return std::isnan(result) ? 0 : result;
 	}
-	float result = std::powf(val, numbers::inv_e_v<float>) * (max - min) + min;
-	return std::isnan( result ) ? 0 : result;
+	float result = std::pow(val, numbers::inv_e_v<float>) * (max - min) + min;
+	return std::isnan(result) ? 0 : result;
 }
 
 inline float fastPow10f(float x)
